STAC example case
