--- examples/STAC/R0.inf
@@ -1,11 +1,0 @@
-!======  ANSYS BladeEditor Export  ========
-Axis of Rotation: Z
-Number of Blade Sets: 39
-Number of Blades Per Set: 1
-Geometry Units: M
-Coordinate System Orientation: Righthanded
-Hub Data File: R0/TS-5_hub.crv
-Shroud Data File: R0/TS-5_shroud.crv
-Profile Data File: R0/TS-5_profile.crv
-Blade Loft Direction: Spanwise
-
