--- examples/STAC/S1.inf
@@ -1,11 +1,0 @@
-!======  ANSYS BladeEditor Export  ========
-Axis of Rotation: Z
-Number of Blade Sets: 39
-Number of Blades Per Set: 1
-Geometry Units: M
-Coordinate System Orientation: Righthanded
-Hub Data File: S1/TS-1_hub.crv
-Shroud Data File: S1/TS-1_shroud.crv
-Profile Data File: S1/TS-1_profile.crv
-Blade Loft Direction: Spanwise
-
