+++ tests/STAC/STAC.TGMachine
@@ -1,0 +1,17 @@
+#  General Comments : Number of Blade Rows must be specified. Init Method can only be inf for now. Blade Rows must point to inf files.
+#  Interface Method : Either Neighbors or Fully Extend. Neighbors is to be used when the curve files specified in the inf files span the entire hub and shroud,
+#                     meaning that they should be trimmed based off of the blade profile locations.
+#                     Note that the order of the inf files in Neighbors mode defines the blade set order, and will be used to determine the adjacency relationship.
+#                     Fully Extend is useful for cases where the hub and shroud curves span only exactly the blade row's mandated endwall sections.
+
+{
+    "Number of Blade Rows": 4,
+    "Init Method": "inf",
+    "Interface Method": "Fully Extend",
+    "Blade Rows": [
+        "S1.inf",
+        "R1.inf",
+        "S0.inf",
+        "R0.inf"
+    ]
+}
